This speedup is on one test suite alone, which was among the most expensive in this regard: 3.8x

```
Benchmark 1: bundle exec rspec ./spec/prog/vnet/cert_nexus_spec.rb
  Time (mean ± σ):      1.539 s ±  0.096 s    [User: 1.183 s, System: 0.174 s]
  Range (min … max):    1.434 s …  1.748 s    10 runs

Benchmark 1: bundle exec rspec ./spec/prog/vnet/cert_nexus_spec.rb
  Time (mean ± σ):      5.739 s ±  1.733 s    [User: 5.396 s, System: 0.181 s]
  Range (min … max):    2.959 s …  8.888 s    10 runs
```

But, it's somewhat bad to be using expensive entropy unnecessarily in tests: it likely be much faster still, with more work, by re-using key material.