@jeremyevans 2024-10-26
Long term, it seems good to switch from RSA to EC. I'm not sure how the public certificate authorities handle EC keys, though.

As you mentioned, this changes production and not just test. For a lesser speedup without a behavior change, we could use lower-bit RSA keys in test and RSA 4096-bit keys in production.  How low we can go depends on the OpenSSL configuration.  The OpenSSL minimum (SSL_MIN_RSA_MODULUS_LENGTH_IN_BYTES) is 512 bit keys (scandalous).  I know the minimum allowed by LibreSSL is 2048. 

Here are the speedups:

```ruby
puts Benchmark.measure{3.times{OpenSSL::PKey::RSA.new(4096)}}
# 4.248120   0.000000   4.248120 (  4.248525)
                                                                    
puts Benchmark.measure{3.times{OpenSSL::PKey::RSA.new(2048)}}
# 0.680071   0.000000   0.680071 (  0.680102)
                                                            
puts Benchmark.measure{3.times{OpenSSL::PKey::RSA.new(1024)}}
# 0.035811   0.000000   0.035811 (  0.035820)
                                                   
 puts Benchmark.measure{3.times{OpenSSL::PKey::RSA.new(512)}}
# 0.018038   0.000000   0.018038 (  0.018034)

```