@fdr 2024-10-27
> Long term, it seems good to switch from RSA to EC. I'm not sure how the public certificate authorities handle EC keys, though.

I think they support them, though it's a bit hard to try both ends of this without a live DNS server to complete the challenge (hello @furkansahin )

https://datatracker.ietf.org/doc/html/rfc8555 mandates P-256 (named as ES256) support at least for account keys. Note acme has both the account key and the customary certificate signing private key that are distinct keys, and could have distinct ciphers, too.

https://wiki.mozilla.org/Security/Server_Side_TLS has been my reference on issues of practical uses of ciphers like this, after many years, In recent years, P-256 has crept to their all-but-lowest/most desperate rung of support.